+++ src/main/java/com/kwang/study/ware/service/WareService.java
@@ -1,5 +1,9 @@
 package com.kwang.study.ware.service;
 
+import com.kwang.study.auth.utils.UserInfoUtils;
+import com.kwang.study.organization.enums.ClassesRoleEnum;
+import com.kwang.study.organization.enums.SchoolRoleEnum;
+import com.kwang.study.organization.pojo.School;
 import com.kwang.study.organization.service.ClassesService;
 import com.kwang.study.auth.mapper.UserMapper;
 import com.kwang.study.organization.pojo.Classes;
@@ -8,6 +12,7 @@ import com.kwang.study.auth.utils.AuthenticationUserUtil;
 import com.kwang.study.enums.FileStorageModuleNameEnum;
 import com.kwang.study.fs.dto.result.*;
 import com.kwang.study.fs.service.FileStorageService;
+import com.kwang.study.organization.service.SchoolService;
 import com.kwang.study.utils.DownloadUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -18,19 +23,25 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.List;
 import java.util.function.Consumer;
+import java.util.stream.Collectors;
 
 @Service
 public class WareService {
     @Autowired
     private FileStorageService fsService;
     @Autowired
-    private UserMapper userMapper;
-    @Autowired
     private ClassesService classesService;
+    @Autowired
+    private SchoolService schoolService;
+    @Autowired
+    private UserInfoUtils userInfoUtils;
 
     @Transactional
     public VoidResult createDirectory(String path) throws IOException {
+        validateWritePermission();
+
         String actualPath = buildActualPath(path);
         return fsService.createDirectory(actualPath);
     }
@@ -41,6 +52,8 @@ public class WareService {
      */
     @Transactional
     public VoidResult createFile(String path, InputStream fileStream, String mimeTypeName) throws IOException {
+        validateWritePermission();
+
         String actualPath = buildActualPath(path);
         return fsService.createFile(actualPath, fileStream, mimeTypeName);
     }
@@ -50,6 +63,8 @@ public class WareService {
      */
     @Transactional
     public VoidResult deleteDirNode(String path) throws IOException {
+        validateWritePermission();
+
         String actualPath = buildActualPath(path);
         return fsService.deleteDirObject(actualPath);
     }
@@ -60,6 +75,8 @@ public class WareService {
      */
     @Transactional
     public VoidResult deleteFileNode(String path) throws IOException {
+        validateWritePermission();
+
         String actualPath = buildActualPath(path);
         return fsService.deleteFileObject(actualPath);
     }
@@ -69,6 +86,8 @@ public class WareService {
      */
     @Transactional
     public VoidResult renameFileNode(String path, String newName) throws IOException {
+        validateWritePermission();
+
         String actualPath = buildActualPath(path);
         return fsService.updateFileObject(actualPath, newName, null, null);
     }
@@ -78,6 +97,8 @@ public class WareService {
      */
     @Transactional
     public VoidResult renameDirNode(String path, String newName) throws IOException {
+        validateWritePermission();
+
         String actualPath = buildActualPath(path);
         return fsService.updateDirObject(actualPath, newName);
     }
@@ -89,12 +110,33 @@ public class WareService {
     public DirObjectResult listDirectoryDetailContents(String path) throws IOException {
         String actualPath = buildActualPath(path);
         DirObjectResult result = fsService.getDirectoryObject(actualPath);
-        // 管理员，班级特殊处理规则，根路径
-        if (AuthenticationUserUtil.currentUserIsAdmin() && "/".equals(path)) {
-            for (DirObjectResult.FileObjectDesc desc : result.getFileObjectDescs()) {
-                Long classId = Long.parseLong(desc.getName());
-                Classes classes = classesService.getClassById(classId);
-                desc.setName(classes.getName());
+        // 管理员，特殊处理规则
+        if (AuthenticationUserUtil.currentUserIsAdmin()) {
+            // 学校
+            if ("/".equals(path)) {
+                List<Long> list = result.getFileObjectDescs().stream()
+                        .map(desc -> Long.parseLong(desc.getName())).collect(Collectors.toList());
+                List<School> schools = schoolService.getBatchSchoolByIds(list);
+                for (int i = 0; i < schools.size(); i++) {
+                    result.getFileObjectDescs().get(i).setName(schools.get(i).getName());
+                }
+            } else if (path.lastIndexOf('/') == 0) {
+                // 班级
+                List<Long> list = result.getFileObjectDescs().stream()
+                        .map(desc -> Long.parseLong(desc.getName())).collect(Collectors.toList());
+                List<Classes> classesList = classesService.getBatchClassByIds(list);
+                for (int i = 0; i < classesList.size(); i++) {
+                    result.getFileObjectDescs().get(i).setName(classesList.get(i).getName());
+                }
+            }
+        }
+        if (userInfoUtils.currentUserInSchoolIsPrincipal() && "/".equals(path)) {
+            // 班级
+            List<Long> list = result.getFileObjectDescs().stream()
+                    .map(desc -> Long.parseLong(desc.getName())).collect(Collectors.toList());
+            List<Classes> classesList = classesService.getBatchClassByIds(list);
+            for (int i = 0; i < classesList.size(); i++) {
+                result.getFileObjectDescs().get(i).setName(classesList.get(i).getName());
             }
         }
         return result;
@@ -126,6 +168,8 @@ public class WareService {
      */
     @Transactional
     public InitMultiUploadResult initMultiUpload(String path, String mimeTypeName) throws IOException {
+        validateWritePermission();
+
         String actualPath = buildActualPath(path);
         return fsService.initMultiUpload(actualPath, mimeTypeName);
     }
@@ -153,31 +197,56 @@ public class WareService {
     // 返回用户该模块的根目录，不以/结尾
     private String buildBasePath() {
         StringBuilder basePath = new StringBuilder(FileStorageModuleNameEnum.WARE_NAME.getModuleName());
-        String userName = AuthenticationUserUtil.getCurrentUserName();
-        User user = userMapper.findByUsernameWithClasses(userName);
+        User user = userInfoUtils.getCurrentUserInfoWithOrgInfo();
         Assert.isTrue(user != null, "没有查找到该用户");
         if (AuthenticationUserUtil.currentUserIsAdmin()) {
             // 管理员能看见all
             return basePath.toString();
         }
-        // 其余只能看见本班级的内容
-        Assert.isTrue(user.getClassMember() != null &&
-                user.getClassMember().getClasses() != null &&
-                user.getClassMember().getClasses().getName() != null, "用户无班级信息");
+        if (user.getSchoolMember() != null && SchoolRoleEnum.PRINCIPAL.getRole()
+                .equals(user.getSchoolMember().getRole())) {
+            // 校长
+            basePath.append('/').append(user.getSchoolMember().getSchool().getId());
+        } else if (user.getClassMember() != null) {
+            // 教师/学生
+            basePath.append('/')
+                    .append(user.getClassMember().getClasses().getSchoolId())
+                    .append('/')
+                    .append(user.getClassMember().getClasses().getId());
+        } else {
+            throw new IllegalArgumentException("用户既无班级信息也无学校信息");
+        }
 
-        basePath.append('/').append(user.getClassMember().getClasses().getId());
         return basePath.toString();
     }
 
     private String buildActualPath(String path) {
-        // 管理员，班级特殊处理规则
+        // 管理员，学校/班级特殊处理规则
         if (AuthenticationUserUtil.currentUserIsAdmin() && !"/".equals(path)) {
             String[] parts = path.split("/");
-            String className = parts[1];
-            Classes classes = classesService.getClassByName(className);
-            Assert.isTrue(classes != null, "路径无班级名称");
+            String schoolName = parts[1];
+            School school = schoolService.getSchoolByName(schoolName);
+            path = path.replaceFirst(schoolName, school.getId().toString());
 
-            path = path.replaceFirst(className, classes.getId().toString());
+            if (parts.length > 2) {
+                String className = parts[2];
+                Classes classes = classesService.getClassByName(className, school.getId());
+                Assert.isTrue(classes != null, "路径无班级名称");
+                path = path.replaceFirst(className, classes.getId().toString());
+            }
+        }
+        // 校长，班级特殊处理规则
+        if (!"/".equals(path)) {
+            User user = userInfoUtils.getCurrentUserInfoWithOrgInfo();
+            // 是校长
+            if (user != null && user.getSchoolMember() != null &&
+                    SchoolRoleEnum.PRINCIPAL.getRole().equals(user.getSchoolMember().getRole())) {
+                String[] parts = path.split("/");
+                String className = parts[1];
+                Classes classes = classesService.getClassByName(className, user.getSchoolMember().getSchoolId());
+                Assert.isTrue(classes != null, "路径无班级名称");
+                path = path.replaceFirst(className, classes.getId().toString());
+            }
         }
 
         String basePath = buildBasePath();
@@ -187,5 +256,16 @@ public class WareService {
             path = path.substring(0, path.length() - 1);
 
         return basePath + path;
+    }
+
+    private void validateWritePermission() {
+        User user = userInfoUtils.getCurrentUserInfoWithOrgInfo();
+        Assert.notNull(user, "无写权限");
+
+        boolean have = AuthenticationUserUtil.currentUserIsAdmin() ||
+                (user.getSchoolMember() != null && SchoolRoleEnum.PRINCIPAL.getRole().equals(user.getSchoolMember().getRole())) ||
+                (user.getClassMember() != null && ClassesRoleEnum.TEACHER.getRole().equals(user.getClassMember().getRole()))
+                ;
+        Assert.isTrue(have, "无写权限");
     }
 }
